F test
Using ANOVA to Test Hypotheses

You use analysis of variance (ANOVA) to test hypotheses about the equality of two or more population means. ANOVA is based on experiments performed on subjects that are independent of each other; in other words, they are not related to each other. 

The following procedure is based on the assumptions that
 
✓ The samples chosen are independent of each other
✓ The underlying populations have equal variances
✓ The populations are normally distributed

Writing the null and alternative hypotheses
 
The null hypothesis is a statement that’s assumed to be true unless you find strong contrary evidence. For testing the hypothesis that three population means are equal, you write:
 
 
In this expression,
H0 represents the null hypothesis
μ1, μ2, μ3 represent the means of population 1, 2 and 3
 
The alternative hypothesis is a statement that you accept in the event that the null hypothesis is rejected (for example, there’s strong evidence against it). When testing the hypothesis that three population means are equal, the alternative hypothesis is simply that the three population means are not equal.
 
This alternative hypothesis can be expressed in different ways, such as:
 
H1: The three means are not all equal
H1: At least one of the three means is different from the others
In these examples, H1 represents the alternative hypothesis.
 
Choosing the level of significance
 
To test a hypothesis, you have to choose a level of significance. The level of significance, designated with α, equals the probability of incorrectly rejecting the null hypothesis when it’s actually true. This is called a Type I error. A Type II error occurs when you fail to reject the null hypothesis when it’s not true. (Check out Chapter 12 for details on Type I and Type II errors in hypothesis testing.)
 
For many business applications, the level of significance is chosen as 0.05 (5 percent). Other frequent choices include 0.001, 0.01, and 0.10.
 
Computing the test statistic
 
The test statistic is a numerical value that is used to determine if the null hypothesis should be rejected. The form of the test statistic depends on the type of hypothesis being tested. If the test statistic has an extremely large positive or negative value, this may be a sign that the null hypothesis is incorrect and should be rejected.
 
Constructing the test statistic for ANOVA is quite complex, compared with other types of hypothesis tests (see Chapter 12 for a discussion of the steps required for hypothesis testing).